+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.11.3] - 2026-06-05
+## [0.12.0] - 2026-06-07
+
+### Added
++ PHP ICS calendar parser johngrogg/ics-parser
++ class.Calendar.php
++ class.Unit.php
++ GET /calendar
 
 ### Changed
 - Security headers